Mandelbrot

Description

Mandelbrot module handles all the calculations with the Complex numbers and the pixels

Documentation

data Complex

simple Complex Number struct

Constructors

C

Fields

re :: Double

im :: Double

type Steps = Int

data ViewWindow

Simple Window Struct

Constructors

View

Fields

upperLeft :: Complex

lowerRight :: Complex

data PictureCoords

init Coords ad size

Constructors

Coords

Fields

x :: Int

y :: Int

data PictureSize

Constructors

Size

Fields

width :: Int

height :: Int

createImage :: ViewWindow -> Steps -> PictureSize -> String -> IO ()

Non parallel version of creating the Image

createImageInParallel :: ViewWindow -> Steps -> PictureSize -> String -> IO ()

Using parrallel processing, this creates the image to display

project :: ViewWindow -> PictureSize -> PictureCoords -> Complex

translates a picture-coords into the ViewWindow

zoomTo :: Double -> Complex -> ViewWindow -> ViewWindow